requirements.txt
torch
torchvision
pyyaml
wandb
pillow
lpips
omegaconf
scikit-image
pandas
